+++ app/config.php
@@ -2,7 +2,7 @@
 
 return [
     'ignore' => [
-        '_partials/*',
-        '_layouts/*'
+        '_layouts/*',
+        '_partials/*'
     ]
 ];